admin ID added
the ID is added to retrieve specific data of the user who logged in on the the website.

+++ include/deletepi.php
@@ -1,12 +1,23 @@
 <?php
 require('db.php');
+
+// Check if the user is logged in
+if (!isset($_SESSION['isUserLoggedIn'])) {
+  echo "<script>window.location.href='login.php';</script>";
+  exit();
+}
+
+$adminId = $_SESSION['admin_id'];
+
 if (isset($_GET['id'])) {
   $id = $_GET['id'];
-  $query = "DELETE FROM personal_info WHERE id=$id";
+
+  // Add a condition to check if the admin_id matches the logged-in user's admin_id
+  $query = "DELETE FROM personal_info WHERE id=$id AND admin_id=$adminId";
   $run = mysqli_query($db, $query);
   if ($run) {
-    echo "<script>window.location.href='../admin/account.php?aboutsetting=true';</script>";
-
+    echo "<script>window.location.href='../Home/account.php?aboutsetting=true';</script>";
+    exit();
   }
 }
 ?>
